Move editor tab to it's own file
