+++ src/renderer/components/EditorToolbar.tsx
@@ -1,14 +1,12 @@
 import {
   faEdit,
-  faFile,
   faPaperclip,
-  faTimes,
   faTrash,
 } from "@fortawesome/free-solid-svg-icons";
 import React, { useCallback, useEffect, useMemo } from "react";
 import styled from "styled-components";
 import { getNoteById, getNoteByPath } from "../../shared/domain/note";
-import { m0, mr2, my2, p2, px2, rounded, THEME } from "../css";
+import { p2, rounded, THEME } from "../css";
 import { Listener, Store } from "../store";
 import { Icon } from "./shared/Icon";
 import { orderBy } from "lodash";
@@ -19,8 +17,8 @@ import { arrayify } from "../../shared/utils";
 import { isProtocolUrl } from "../../shared/domain/protocols";
 import OpenColor from "open-color";
 import { deleteNoteIfConfirmed } from "../utils/deleteNoteIfConfirmed";
+import { EditorTab } from "./EditorTab";
 
-export const EDITOR_TAB_ATTRIBUTE = "data-editor-tab";
 export const TOOLBAR_HEIGHT = "4.3rem"; // 4.2rem + 1px for border
 
 export interface EditorToolbarProps {
@@ -343,134 +341,6 @@ const TabsScrollable = styled(Scrollable)`
   }
 `;
 
-export interface EditorTabProps {
-  noteId: string;
-  noteName: string;
-  active?: boolean;
-  onClick: (noteId: string) => void;
-  onClose: (noteId: string) => void;
-}
-
-export function EditorTab(props: EditorTabProps): JSX.Element {
-  const { noteId, noteName, active } = props;
-
-  const onDeleteClick = (ev: React.MouseEvent<HTMLElement>) => {
-    // Need to stop prop otherwise it'll trigger onClick of tab.
-    ev.stopPropagation();
-    props.onClose(noteId);
-  };
-
-  if (active) {
-    return (
-      <StyledSelectedTab
-        key={noteId}
-        title={noteName}
-        onClick={() => props.onClick(noteId)}
-        {...{ [EDITOR_TAB_ATTRIBUTE]: noteId }}
-      >
-        <FlexRow>
-          <StyledNoteIcon icon={faFile} size="lg" />
-          <StyledSelectedText>{noteName}</StyledSelectedText>
-        </FlexRow>
-        <StyledDelete
-          icon={faTimes}
-          onClick={onDeleteClick}
-          className="delete"
-          title={`Close ${noteName}`}
-        />
-      </StyledSelectedTab>
-    );
-  } else {
-    return (
-      <StyledTab
-        key={noteId}
-        title={noteName}
-        onClick={() => props.onClick(noteId)}
-        {...{ [EDITOR_TAB_ATTRIBUTE]: noteId }}
-      >
-        <FlexRow>
-          <StyledNoteIcon icon={faFile} size="lg" />
-          <StyledText>{noteName}</StyledText>
-        </FlexRow>
-        <StyledDelete
-          icon={faTimes}
-          onClick={onDeleteClick}
-          className="delete"
-          title={`Close ${noteName}`}
-        />
-      </StyledTab>
-    );
-  }
-}
-
-const StyledNoteIcon = styled(Icon)`
-  margin-right: 1rem;
-`;
-
-const FlexRow = styled.div`
-  display: flex;
-  flex-direction: row;
-  align-items: center;
-`;
-
-const StyledTab = styled.a`
-  display: inline-flex;
-  align-items: center;
-  justify-content: space-between;
-  width: 12rem;
-  ${px2}
-  ${my2}
-  ${mr2}
-
-  border-radius: 0.4rem;
-  height: 2.6rem;
-
-  &:hover {
-    background-color: ${THEME.editor.tabs.hoveredTabBackground};
-
-    .delete {
-      display: block;
-    }
-  }
-`;
-
-const StyledSelectedTab = styled(StyledTab)`
-  background-color: ${THEME.editor.tabs.activeTabBackground};
-
-  .delete {
-    display: block;
-  }
-`;
-
-const StyledText = styled.div`
-  font-size: 1.2rem;
-  font-weight: 500;
-  text-overflow: ellipsis;
-  white-space: nowrap;
-  overflow: hidden;
-`;
-
-const StyledSelectedText = styled(StyledText)`
-  font-size: 1.2rem;
-  font-weight: bold;
-  color: ${THEME.editor.tabs.activeTabFont};
-  text-overflow: ellipsis;
-  white-space: nowrap;
-  overflow: hidden;
-`;
-
-const StyledDelete = styled(Icon)`
-  border-radius: 0.4rem;
-  ${p2}
-  ${m0}
-
-  &:hover {
-    cursor: pointer;
-    color: ${THEME.editor.tabs.deleteColor};
-    background-color: ${THEME.editor.tabs.deleteHoverBackground};
-  }
-`;
-
 export const openTab: Listener<"editor.openTab"> = async (ev, ctx) => {
   // Keep in sync with sidebar.openSelectedNotes listener
   const { editor, notes } = ctx.getState();
@@ -559,12 +429,3 @@ export const updateTabsScroll: Listener<"editor.updateTabsScroll"> = async (
     });
   }
 };
-
-export function getEditorTabAttribute(element: HTMLElement): string | null {
-  const parent = element.closest(`[${EDITOR_TAB_ATTRIBUTE}]`);
-  if (parent != null) {
-    return parent.getAttribute(EDITOR_TAB_ATTRIBUTE);
-  }
-
-  return null;
-}
