+++ src/renderer/menus/contextMenu.ts
@@ -13,7 +13,7 @@ import { getSidebarMenuAttribute } from "../components/SidebarMenu";
 import { getShortcutLabels } from "../io/shortcuts";
 import { Store } from "../store";
 import { Section } from "../../shared/ui/app";
-import { getEditorTabAttribute } from "../components/EditorToolbar";
+import { getEditorTabAttribute } from "../components/EditorTab";
 import { IpcChannel } from "../../shared/ipc";
 import { Config } from "../../shared/domain/config";
